Updated createRetina_OCL implementation to use the declared double instead of float parameters

+++ modules/bioinspired/include/opencv2/bioinspired/retina.hpp
@@ -449,7 +449,7 @@ CV_EXPORTS_W Ptr<Retina> createRetina(Size inputSize, const bool colorMode, int 
 
 #ifdef HAVE_OPENCL
 CV_EXPORTS_W Ptr<Retina> createRetina_OCL(Size inputSize);
-CV_EXPORTS_W Ptr<Retina> createRetina_OCL(Size inputSize, const bool colorMode, int colorSamplingMethod = RETINA_COLOR_BAYER, const bool useRetinaLogSampling = false, const float reductionFactor = 1.0f, const float samplingStrenght = 10.0f);
+CV_EXPORTS_W Ptr<Retina> createRetina_OCL(Size inputSize, const bool colorMode, int colorSamplingMethod = RETINA_COLOR_BAYER, const bool useRetinaLogSampling = false, const double reductionFactor = 1.0f, const double samplingStrenght = 10.0f);
 #endif
 
 //! @}
